Rename modules pythonically and add module index
