+++ copulas/multivariate/__init__.py
@@ -1,0 +1,5 @@
+from copulas.multivariate.base import Multivariate
+from copulas.multivariate.gaussian import GaussianMultivariate
+from copulas.multivariate.vine import Tree, Vine
+
+__all__ = ('Multivariate', 'GaussianMultivariate', 'Vine', 'Tree')
