+++ copulas/univariate/__init__.py
@@ -1,0 +1,5 @@
+from copulas.univariate.base import Univariate
+from copulas.univariate.gaussian import GaussianUnivariate
+from copulas.univariate.kde import KDEUnivariate
+
+__all__ = ('Univariate', 'GaussianUnivariate', 'KDEUnivariate')
