+++ copulas/univariate/gaussian.py
@@ -3,12 +3,12 @@ import logging
 import numpy as np
 from scipy.stats import norm
 
-from copulas.univariate.UnivariateDistrib import UnivariateDistrib
+from copulas.univariate.base import Univariate
 
 LOGGER = logging.getLogger(__name__)
 
 
-class GaussianUnivariate(UnivariateDistrib):
+class GaussianUnivariate(Univariate):
     """ Gaussian univariate model """
 
     def __init__(self):
