+++ tests/univariate/test_kde.py
@@ -8,10 +8,11 @@ from unittest import TestCase
 import numpy as np
 import scipy
 
-from copulas.univariate.KDEUnivariate import KDEUnivariate
+from copulas.univariate import KDEUnivariate
 
 
 class TestKDEUnivariate(TestCase):
+
     def setup_norm(self):
         """set up the model to fit standard norm data"""
         self.kde = KDEUnivariate()
